Added multiple transient data gathering
Added inside the 'voeventDB_Connector' the ability to allow the script to gather multiple transients as compared to one before.  It starts gathering from 8/01/2016, but can be modified to be earlier and also a certain limit as to the latest 10 possibly.

Also removed the 'recieved-on' variable as there was no need for it.